move getLocalPing to PingCommand
